third_party/WebKit/LayoutTests/crypto/subtle/rsa-oaep/encrypt-failures.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/keys.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests bad algorithm inputs for RSA-OAEP encrypt");

jsTestIsAsync = true;

var data = new Uint8Array(16);

function importPublicKey()
{
    var usages = ['encrypt'];
    var extractable = true;

    var importData = hexStringToUint8Array(kKeyData.rsa1.spki);
    var importAlgorithm = { name: 'rsa-oaep', hash: {name: 'SHA-256'} };

    return crypto.subtle.importKey('spki', importData, importAlgorithm, extractable, usages);
}

Promise.resolve(null).then(function(result) {
    return importPublicKey();
}).then(function(result) {
    publicKey = result;

    shouldEvaluateAs("publicKey.algorithm.name", "RSA-OAEP");
    shouldEvaluateAs("publicKey.algorithm.hash.name", "SHA-256");
    shouldEvaluateAs("publicKey.type", "public");
    shouldEvaluateAs("publicKey.extractable", true);
    shouldEvaluateAs("publicKey.usages.join(',')", "encrypt");

    return crypto.subtle.encrypt({name: "rsa-oaep", label: -1}, publicKey, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.encrypt({name: "rsa-oaep", label: "foo"}, publicKey, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
