third_party/WebKit/LayoutTests/crypto/subtle/rsa-oaep/encrypt-failures-expected.txt

Tests bad algorithm inputs for RSA-OAEP encrypt

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS publicKey.algorithm.name is "RSA-OAEP"
PASS publicKey.algorithm.hash.name is "SHA-256"
PASS publicKey.type is "public"
PASS typeof publicKey.extractable is 'boolean'
PASS publicKey.extractable is true
PASS publicKey.usages.join(',') is "encrypt"
error is: TypeError: RsaOaepParams: label: Not a BufferSource
error is: TypeError: RsaOaepParams: label: Not a BufferSource
PASS successfullyParsed is true

TEST COMPLETE
